+++ demo/meeting/src/containers/Navigation/index.tsx
@@ -9,8 +9,6 @@ import {
   NavbarItem,
   Attendees,
   Eye,
-  Information,
-  Flex,
   SignalStrength
 } from 'amazon-chime-sdk-component-library-react';
 
@@ -19,42 +17,29 @@ import { useAppState } from '../../providers/AppStateProvider';
 import LocalMediaStreamMetrics from '../LocalMediaStreamMetrics';
 
 const Navigation = () => {
-  const { toggleRoster, toggleMetrics, closeNavbar } = useNavigation();
+  const { toggleRoster, closeNavbar } = useNavigation();
   const { theme, toggleTheme } = useAppState();
 
   return (
     <Navbar className="nav" flexDirection="column" container>
       <NavbarHeader title="Navigation" onClose={closeNavbar} />
-      <Flex
-        css={`
-          margin-top: 0rem;
-        `}
+      <NavbarItem
+        icon={<Attendees />}
+        onClick={toggleRoster}
+        label="Attendees"
+      />
+      <NavbarItem
+        icon={<Eye />}
+        onClick={toggleTheme}
+        label={theme === 'light' ? 'Dark mode' : 'Light mode'}
+      />
+      <NavbarItem
+        icon={<SignalStrength />}
+        onClick={() => {}}
+        label="Media metrics"
       >
-        <NavbarItem
-          icon={<Attendees />}
-          onClick={toggleRoster}
-          label="Attendees"
-        />
-        <NavbarItem
-          icon={<Eye />}
-          onClick={toggleTheme}
-          label={theme === 'light' ? 'Dark mode' : 'Light mode'}
-        />
-        <NavbarItem
-          icon={<Information />}
-          onClick={toggleMetrics}
-          label="Bandwidth metrics"
-        />
-      </Flex>
-      <Flex marginTop="auto">
-        <NavbarItem
-          icon={<SignalStrength />}
-          onClick={() => {}}
-          label="Meeting metrics"
-        >
-          <LocalMediaStreamMetrics />
-        </NavbarItem>
-      </Flex>
+        <LocalMediaStreamMetrics />
+      </NavbarItem>
     </Navbar>
   );
 };