+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Doc] Add documentaion in introduction on how to use `MeetingSessionPostLogger` to post
   Amazon Chime JS SDK logs.
-- Add useMediaStreamMetrics hook to expose audio, video and bandwidth data.
+- Add `useMediaStreamMetrics` hook to expose audio, video and bandwidth data.
 
 ### Changed